Back study member

See merge request s06-webmobile2-sub2/S06P12D207!19

+++ backend-java/src/main/java/com/web/curation/service/MatchService.java
@@ -16,6 +16,7 @@ import org.springframework.web.server.ResponseStatusException;
 import com.web.curation.dto.match.MatchArticleRequest;
 import com.web.curation.dto.match.MatchArticleResponse;
 import com.web.curation.dto.match.MatchJoinUserResponse;
+import com.web.curation.model.match.JoinState;
 import com.web.curation.model.account.MyUser;
 import com.web.curation.model.match.MatchArticle;
 import com.web.curation.model.match.MatchJoin;
@@ -54,9 +55,9 @@ public class MatchService {
     }
     
     public MatchArticleResponse getArticle(Long articleId) {
-    	MatchArticle article = articleRepo.findById(articleId)
-    			.orElseThrow(() -> new IllegalStateException(
-    					"article with id " + articleId + " does not exist"));
+    	MatchArticle article = articleRepo.findById(articleId).orElseThrow(() -> new ResponseStatusException(
+																HttpStatus.NOT_FOUND, "존재하지 않는 게시글 id입니다.",
+																new IllegalArgumentException()));
     	MatchArticleResponse response = article.toResponse();
     	return response;
     }
@@ -67,8 +68,7 @@ public class MatchService {
     	MatchArticle article = articleForm.toEntity();
     	MyUser writer = userRepo.findById(writerId)
     			.orElseThrow(() -> new ResponseStatusException(
-							HttpStatus.BAD_REQUEST,
-							"존재하지 않는 유저 id입니다.",
+							HttpStatus.NOT_FOUND, "존재하지 않는 유저 id입니다.",
 							new IllegalArgumentException()));
     	article.setWriter(writer);
 //    	writer.getMatchArticles().add(article);
@@ -79,16 +79,20 @@ public class MatchService {
     @Transactional // 변경된 데이터를 DB에 저장
     public void updateArticle(
     		Long articleId,
+    		Long userId,
     		String title,
     		String category,
     		String content,
     		Boolean state) {
     	MatchArticle article = articleRepo.findById(articleId)
     			.orElseThrow(() -> new ResponseStatusException(
-						HttpStatus.BAD_REQUEST,
-						"존재하지 않는 게시글 id입니다.",
+						HttpStatus.NOT_FOUND, "존재하지 않는 게시글 id입니다.",
 						new IllegalArgumentException()));
-    	
+    	if (article.getWriter().getId() != userId) {
+    		throw new ResponseStatusException(
+					HttpStatus.BAD_REQUEST, "게시글(댓글) 작성자가 아닙니다.",
+					new IllegalArgumentException());
+    	}
     	if (title != null && title.length() > 0 && !Objects.equals(article.getTitle(), title)) {
     		article.setTitle(title);
     	}
@@ -106,7 +110,15 @@ public class MatchService {
     	}
     }
     
-    public void deleteArticle(Long articleId) {
+    public void deleteArticle(Long articleId, Long userId) {
+    	MatchArticle article = articleRepo.findById(articleId).orElseThrow(() -> new ResponseStatusException(
+																HttpStatus.NOT_FOUND, "존재하지 않는 게시글 id입니다.",
+																new IllegalArgumentException()));
+    	if (article.getWriter().getId() != userId) {
+    		throw new ResponseStatusException(
+					HttpStatus.BAD_REQUEST, "게시글(댓글) 작성자가 아닙니다.",
+					new IllegalArgumentException());
+    	}
     	articleRepo.deleteById(articleId);
     }
     
@@ -124,22 +136,26 @@ public class MatchService {
     
     // 스터디 신청
     public void joinStudy(Long articleId, Long joinUserId, String content) {
-    	MatchArticle article = articleRepo.findById(articleId).get();
+    	MatchArticle article = articleRepo.findById(articleId).orElseThrow(() -> new ResponseStatusException(
+																HttpStatus.NOT_FOUND, "존재하지 않는 게시글 id입니다.",
+																new IllegalArgumentException()));
     	MyUser joinUser = userRepo.findById(joinUserId)
     			.orElseThrow(() -> new ResponseStatusException(
-						HttpStatus.BAD_REQUEST,
-						"존재하지 않는 유저 id입니다.",
+						HttpStatus.NOT_FOUND, "존재하지 않는 유저 id입니다.",
 						new IllegalArgumentException()));
+    	
 
     	if (article.getWriter().getId() == joinUserId) {
     		throw new ResponseStatusException(
-    				HttpStatus.BAD_REQUEST,
-    				"작성한 모집글에 신청할 수 없습니다.",
+    				HttpStatus.BAD_REQUEST, "작성한 모집글에 신청할 수 없습니다.",
+    				new IllegalArgumentException());
+    	} else if (studyJoinRepo.findByJoinMemberIdAndJoinStudyId(joinUserId, article.getStudyId()).isPresent()) {
+    		throw new ResponseStatusException(
+    				HttpStatus.BAD_REQUEST, "스터디 멤버는 신청할 수 없습니다.",
     				new IllegalArgumentException());
     	} else if (joinRepo.findByJoinUserIdAndJoinArticleId(joinUserId, articleId).isPresent()) {
     		throw new ResponseStatusException(
-    				HttpStatus.BAD_REQUEST,
-    				"이미 신청한 모집글입니다.",
+    				HttpStatus.BAD_REQUEST, "이미 신청한 모집글입니다.",
     				new IllegalArgumentException());
     	} else {
         	MatchJoin join = new MatchJoin();
@@ -170,17 +186,17 @@ public class MatchService {
 	// 모집글에서 '승인'으로 멤버 추가
 	@Transactional
 	public void addNewMatchMember(Long articleId, Long joinUserId) {
-		MatchArticle article = articleRepo.findById(articleId).get();
+		MatchArticle article = articleRepo.findById(articleId).orElseThrow(() -> new ResponseStatusException(
+																HttpStatus.NOT_FOUND, "존재하지 않는 게시글 id입니다.",
+																new IllegalArgumentException()));
 
 		MyUser leader = article.getWriter();
 		MyUser joinUser = userRepo.findById(joinUserId).orElseThrow(() -> new ResponseStatusException(
-													HttpStatus.BAD_REQUEST,
-													"존재하지 않는 유저 id입니다.",
+													HttpStatus.NOT_FOUND, "존재하지 않는 유저 id입니다.",
 													new IllegalArgumentException()));
 		if (leader.getId() == joinUserId) {
     		throw new ResponseStatusException(
-    				HttpStatus.BAD_REQUEST,
-    				"모집글 작성자는 일반 회원으로 추가할 수 없습니다.",
+    				HttpStatus.BAD_REQUEST, "모집글 작성자는 일반 회원으로 추가할 수 없습니다.",
     				new IllegalArgumentException());
 		}
 
@@ -188,7 +204,7 @@ public class MatchService {
 		if (article.getStudyId() == null) {
 			// 스터디 생성
 			Study study = new Study();
-			study.setLeader(leader);
+//			study.setLeader(leader);
 			Study saved = studyRepo.save(study);
 			
 			// 스터디에 유저추가(리더)
@@ -197,22 +213,50 @@ public class MatchService {
 			// 스터디에 유저추가(멤버1)
 			StudyJoin join2 = new StudyJoin(null, joinUser, saved, false);
 			studyJoinRepo.save(join2);
+			MatchJoin matchJoin = joinRepo.findByJoinUserIdAndJoinArticleId(joinUserId, articleId)
+											.orElseThrow(() -> new ResponseStatusException(
+																HttpStatus.BAD_REQUEST, "유저 또는 게시글 id를 확인하세요",
+																new IllegalArgumentException()));
+			matchJoin.setState(JoinState.APPROVED);
 			// matchArticle에 추가
 			article.setStudyId(saved.getId());
 		} else {
 			Long studyId = article.getStudyId();
-			Study study = studyRepo.findById(studyId).get();
+			Study study = studyRepo.findById(studyId).orElseThrow(() -> new ResponseStatusException(
+														HttpStatus.BAD_REQUEST, "모집글의 스터디 정보를 알 수 없습니다.",
+														new IllegalArgumentException()));
 			if (studyJoinRepo.findByJoinMemberIdAndJoinStudyId(joinUserId, studyId).isPresent()) {
 	    		throw new ResponseStatusException(
-	    				HttpStatus.BAD_REQUEST,
-	    				"이미 스터디에 가입한 회원입니다.",
+	    				HttpStatus.BAD_REQUEST, "이미 스터디에 가입한 회원입니다.",
 	    				new IllegalArgumentException());
 			} else {
 				// 스터디에 유저추가(멤버1)
 				StudyJoin join = new StudyJoin(null, joinUser, study, false);
 				studyJoinRepo.save(join);
+				MatchJoin matchJoin = joinRepo.findByJoinUserIdAndJoinArticleId(joinUserId, articleId)
+												.orElseThrow(() -> new ResponseStatusException(
+																	HttpStatus.BAD_REQUEST, "유저 또는 게시글 id를 확인하세요",
+																	new IllegalArgumentException()));
+				matchJoin.setState(JoinState.APPROVED);
 			}
 		}
 
+	}
+	
+	// 신청자 거부
+	@Transactional
+	public void denyJoinUser(Long articleId, Long joinUserId) {
+		MatchJoin join = joinRepo.findByJoinUserIdAndJoinArticleId(joinUserId, articleId)
+												.orElseThrow(() -> new ResponseStatusException(
+														HttpStatus.BAD_REQUEST, "유저 또는 게시글 id를 확인하세요",
+														new IllegalArgumentException()));
+		if (join.getState() == null) {
+			join.setState(JoinState.DENIED);
+		} else {
+			throw new ResponseStatusException(
+    				HttpStatus.BAD_REQUEST, "승인 또는 거부한 신청자입니다.",
+    				new IllegalArgumentException());
+		}
+		
 	}
 }
